misc/HtmlTestSuite/results/test-1021.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>HTML Test: base_multiple</title>
<link href="http://www.intel.com/" rel="author" title="Intel"/>
<link href="https://html.spec.whatwg.org/multipage/#the-base-element" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body>
  <div id="log"/>
  <iframe frameborder="0" id="test1" src="example.html" style="width:0;height:0"/>
  <iframe frameborder="0" id="test2" name="targetWin" src="example.html" style="width:0;height:0"/>
  <script>
    async_test(function() {
      window.onload = this.step_func(function() {
        var fr1 = document.getElementById("test1");
        fr1.addEventListener("load", this.unreached_func("loaded in the wrong iframe"));

        var fr2 = document.getElementById("test2");
        fr2.addEventListener("load", this.step_func_done(function () {
          var doc2 = fr2.contentDocument;
          assert_not_equals(doc2.location.href.indexOf("example2.html"), -1, "The target attribute does not impact the a element.");
          assert_equals(doc2.getElementById("d1").innerHTML, "PASS", "The opend page should be the example2.html.");
        }), true);

        fr1.contentDocument.getElementById("a1").click();
      });
    }, "The attributes of the a element must be affected by the first base element");
  </script>
</body>
</html>
